server/requirements.txt
Django
djangorestframework
numpy
pandas
aiohttp
daphne
django-cors-headers
